+++ priv/repo/migrations/20200114093407_add_group_to_templates_content.exs
@@ -98,5 +98,4 @@ defmodule TdDf.Repo.Migrations.AddGroupToTemplatesContent do
       )
     )
   end
-
 end